# openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/user": {
      "post": {
        "description": "Jammit 계정 생성",
        "operationId": "JammitUserController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "email": "user@example.com",
                  "password": "password123",
                  "username": "jammituser",
                  "nickname": "닉네임",
                  "preferredGenres": [
                    "ROCK",
                    "POP"
                  ],
                  "preferredBandSessions": [
                    "VOCAL",
                    "ELECTRIC_GUITAR"
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "가입 성공 (CommonResponseDto)"
          },
          "400": {
            "description": "이메일 중복 등"
          }
        },
        "summary": "회원가입",
        "tags": [
          "Jammit User"
        ]
      },
      "get": {
        "description": "JWT 기준 로그인 유저 정보",
        "operationId": "JammitUserController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "프로필 객체 (CommonResponseDto)"
          },
          "401": {
            "description": "미인증"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "내 프로필 조회",
        "tags": [
          "Jammit User"
        ]
      },
      "put": {
        "operationId": "JammitUserController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "email": "new@example.com",
                  "username": "newname",
                  "password": "newpassword",
                  "preferredGenres": [
                    "JAZZ"
                  ],
                  "preferredBandSessions": [
                    "DRUM"
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "수정 후 프로필 (CommonResponseDto)"
          },
          "400": {
            "description": "유저 없음 등"
          },
          "401": {
            "description": "미인증"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "내 프로필 수정",
        "tags": [
          "Jammit User"
        ]
      }
    },
    "/user/exists": {
      "get": {
        "description": "회원가입 전 사용 가능 여부",
        "operationId": "JammitUserController_exists",
        "parameters": [
          {
            "name": "email",
            "required": true,
            "in": "query",
            "schema": {
              "example": "user@example.com",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{ exists: boolean } 형태 (CommonResponseDto)"
          }
        },
        "summary": "이메일 중복 확인",
        "tags": [
          "Jammit User"
        ]
      }
    },
    "/user/image": {
      "put": {
        "description": "직접 URL을 저장할 때 사용",
        "operationId": "JammitUserController_updateImage",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "orgFileName": "photo.jpg",
                  "profileImagePath": "https://example.com/profile/photo.jpg"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "갱신된 프로필 (CommonResponseDto)"
          },
          "401": {
            "description": "미인증"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "프로필 이미지 URL 갱신",
        "tags": [
          "Jammit User"
        ]
      }
    },
    "/user/{userId}/profile-image": {
      "post": {
        "description": "multipart file → 스토리지 업로드 후 URL 반환",
        "operationId": "JammitUserController_uploadProfile",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "본인 users.id 와 일치해야 함",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "이미지 파일"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "업로드된 이미지 URL (CommonResponseDto)"
          },
          "400": {
            "description": "파일 없음·유저 없음"
          },
          "401": {
            "description": "미인증"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "프로필 이미지 업로드",
        "tags": [
          "Jammit User"
        ]
      }
    },
...
  },
  "info": {
    "title": "jammit api문서",
    "description": "jammit 개발을 위한 API 문서입니다.\n\n🔗 OpenAPI 명세 다운로드: /openapi-spec.json",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "JWT access token",
        "in": "header"
      }
    },
...
  }
}